requirements.txt
flask
flask-sqlalchemy
flask-migrate
pillow
flask-seeder
